implementing normalize

+++ cpp_cuda_c/src/spline_psf_gpu.cu
@@ -524,13 +524,14 @@ auto kernel_roi(spline *sp, float *rois, const int npx, const int npy, const flo
     float phot = phot_[r];
 
     /* Compute delta. Will be the same for all following px */
-    x0 = (int)floor(xc);
+    // floats use floorf I think, stolen from another branch
+    x0 = (int)floorf(xc);
     x_delta = xc - x0;
 
-    y0 = (int)floor(yc);
+    y0 = (int)floorf(yc);
     y_delta = yc - y0;
 
-    z0 = (int)floor(zc);
+    z0 = (int)floorf(zc);
     z_delta = zc - z0;
 
     int n_threads = min(1024, npx * npy);  // max number of threads per block
@@ -557,13 +558,13 @@ auto kernel_derivative_roi(spline *sp, float *rois, float *drv_rois, const int n
     float bg = bg_[r];
 
     /* Compute delta. Will be the same for all following px */
-    x0 = (int)floor(xc);
+    x0 = (int)floorf(xc);
     x_delta = xc - x0;
 
-    y0 = (int)floor(yc);
+    y0 = (int)floorf(yc);
     y_delta = yc - y0;
 
-    z0 = (int)floor(zc);
+    z0 = (int)floorf(zc);
     z_delta = zc - z0;
 
     int n_threads = min(1024, npx * npy);  // max number of threads per block
